Null check data before stringifying/serializing and check validity
The BIP and AVRCP-BIP specifications claim you must (1) have a fixed
version, (2) have an image handle and (3) have an image format that
captures the imaging thumbnail format defined in 5.14.2.2.1 of AVRCP
1.6.

This change makes it so
(1) we no longer assume the data we get will be completely in spec.
We null check before serializing what we've built or printing what
we received for debug purposes
(2) We can check the BipImageProperties object for validity against the
specification requirements.

Significant tests have been added to show the new code works and is
robust.

Tag: #stability
Bug: 181811272
Test: atest BluetoothInstrumentationTests
Test: Interop with Pixel, iPhone, and Samsung devices
Change-Id: I4340de3311551372c847b79b6ee5f49c7adc2a9a